chore(mcp): move insight visualization tests into the tests folder
Relocate the consolidated suite from src/ui-apps/components/ to the
standard tests/ layout (tests/unit/ + tests/fixtures/), adopting the
folder's conventions: kebab-case file names and @/ path-alias imports.
This also reverts the vitest include tweak — the unit project's
tests/** glob already picks the suite up.

Generated-By: PostHog Code
Task-Id: 8a135279-2519-48dd-ba25-2d52eaa39b0d
